+++ src/index.ts
@@ -1,0 +1,5 @@
+import Server from "./main/Server";
+
+const Simpler = Server
+
+export default Simpler