+++ src/main/Router/router.interface.ts
@@ -1,0 +1,12 @@
+interface IRouter {
+  addRoute: AddRouteFunction;
+  addRoutes: AddRoutesFunction;
+  getRoute: GetRouteFunction;
+  getPathVariables: GetPathVariablesFunction;
+  getQueryParams: GetQueryParamsFunction;
+  addStaticDirectory: AddStaticDirectoryFunction;
+  addStaticDirectories: AddStaticDirectoriesFunction;
+  getStaticDirs: GetStaticDirsFunction;
+}
+
+export default IRouter;