+++ src/main/Server/server.interface.ts
@@ -1,0 +1,7 @@
+interface IServer {
+  handleRequest: HandleRequestFunction;
+  listen: ListenFunction;
+  response: ResponseFunction;
+}
+
+export default IServer;